+++ src/Projectile.java
@@ -8,6 +8,6 @@ public interface Projectile {    // look into angles
     public void move();
     public void hit(); 
     public void draw(Graphics2D g2);
-    public void spawn(int x, int y, double angle);
+    public void spawn(double x, double y, double angle);
 
 }
